string hashing done I think

+++ StudentCode/IntHash.h
@@ -6,14 +6,14 @@ template <class T>
 class IntHash :public HashTable<int,T>
 {
 private:
-	int h1(int k);
-	int h2(int k);
+	uint64_t h1(int k);
+	uint64_t h2(int k);
 public:
 	IntHash(int m = 10) :HashTable<int,T>(m) {}
 };
 
 template <class T>
-int IntHash<T>::h1(int k)
+uint64_t IntHash<T>::h1(int k)
 {
 
 	return k % this->size
@@ -21,7 +21,7 @@ int IntHash<T>::h1(int k)
 }
 
 template <class T>
-int IntHash<T>::h2(int k)
+uint64_t IntHash<T>::h2(int k)
 {
 	return 1 + k % (this->size - 1)
 }
